replace  tooltip view with ui kit component
