--- src/features/tooltip/Tooltip/Tooltip.module.css
@@ -1,25 +1,0 @@
-@custom-media --mobile (max-width: 960px);
-/* var(--md) */
-
-.tooltipContainer {
-  position: absolute;
-  left: 0;
-  top: 0;
-  width: 100%;
-  height: 100%;
-  z-index: var(--tooltip);
-  display: flex;
-  flex-direction: row-reverse;
-}
-
-.tooltipContainer.hoverTooltip {
-  pointer-events: none;
-
-  & .popup {
-    pointer-events: visible;
-  }
-}
-
-.unset {
-  display: none;
-}
