--- src/features/tooltip/TooltipContent/TooltipContent.module.css
@@ -1,80 +1,0 @@
-.tooltipContent {
-  position: absolute;
-  display: flex;
-  padding: var(--unit) var(--unit) var(--half-unit) var(--unit);
-  max-width: 268px;
-  max-height: 200px;
-  overflow-y: auto;
-  background-color: var(--base-strong);
-  color: var(--base-weak);
-  border-radius: var(--border-radius);
-}
-
-.closeIcon {
-  position: sticky;
-  top: 0;
-  cursor: pointer;
-  height: 16px;
-  display: flex;
-  color: var(--faint-strong);
-}
-
-.closeIcon svg {
-  width: var(--double-unit);
-  height: var(--double-unit);
-}
-
-.markdown {
-  & p {
-    margin: 0;
-  }
-  & a {
-    word-break: break-all;
-    &:visited,
-    &:link,
-    &:active,
-    &:hover {
-      color: var(--accent-strong);
-    }
-  }
-}
-
-@media (--mobile) {
-  .closeIcon {
-    cursor: pointer;
-    height: 10px;
-    display: flex;
-    color: var(--base-strong);
-  }
-
-  .markdown {
-    flex: 1 1 0;
-    & a {
-      &:visited,
-      &:link,
-      &:active,
-      &:hover {
-        color: var(--accent-strong);
-      }
-    }
-  }
-}
-
-@media (--mobile) {
-  .tooltipContent {
-    position: absolute;
-    display: flex;
-    z-index: var(--tooltip);
-    left: 0;
-    bottom: calc(0px - var(--half-unit));
-    width: 100%;
-    padding: var(--double-unit);
-    max-height: 90%;
-    max-width: 100%;
-    background-color: var(--base-weak);
-    color: var(--base-strong);
-    font-weight: 400;
-    font-size: 16px;
-    border-radius: var(--unit) var(--unit) 0 0;
-  }
-}
