--- src/features/tooltip/types.ts
@@ -1,3 +1,0 @@
-export type PlacementFn = (coords: { x: number; y: number }) => TooltipPlacement;
-export type TooltipCoords = { x: number; y: number };
-export type TooltipPlacement = 'top-left' | 'top-right' | 'bottom-left' | 'bottom-right';
